requirements.txt
google-generativeai
python-dotenv
sentence-transformers
faiss-cpu
pypdf
numpy
requests
